requirements.txt
Pillow
argparse
os
pytest